Add brand filter dropdown to advanced search
Added a dropdown (`MudSelect`) for selecting a brand in the advanced
search section of the `Products.razor` file, allowing users to filter
products by brand with an "All Brands" option. Updated the search
function in `Products.razor.cs` to include the selected brand ID
(`SearchBrandId`). Changed the type of `SearchBrandId` from `Guid` to
`Guid?` to support the nullable brand ID for the "All Brands" option.